# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zakat Fintech API Documentation",
        "description": "A description of Zakat Fintech API",
        "contact": {
            "email": "firdausishaddin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://zakat-fintech.test",
            "description": "Local Environment"
        }
    ],
    "paths": {
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get list of Services",
                "description": "Retrieves a list of all available services.",
                "operationId": "7eb9058f0d2f3bc0a9a22df5d47b1f66",
                "responses": {
                    "200": {
                        "description": "List of services retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Service"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Store a Service into database",
                "description": "Create a new service entry in the database.",
                "operationId": "457b0baebe6cc22e98acf5126b2dbeb4",
                "requestBody": {
                    "description": "Data required to create a new service",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bantuan Makanan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get a Service by ID",
                "description": "Retrieves detailed information about a specific service using its ID.",
                "operationId": "5b45d934dcfdf987bf09e18588b656d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier of the service",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update a specific service",
                "description": "Updates the details of a specific service by ID.",
                "operationId": "f59e1d8f9cf9e494e99f9490cf609a1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Object containing updated service information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Service"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a Service by ID",
                "description": "Deletes a specific service identified by its ID.",
                "operationId": "a439d1bfff812f8f6cf7c948310a9f50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique identifier of the service to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Service successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Service": {
                "title": "Service",
                "description": "References Service Model",
                "properties": {
                    "name": {
                        "description": "Name of the service",
                        "type": "string",
                        "example": "Bantuan Makanan"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://zakat-fintech.test/oauth/authorize",
                        "tokenUrl": "https://zakat-fintech.test/oauth/token",
                        "refreshUrl": "https://zakat-fintech.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Services",
            "description": "API Endpoints of Services"
        }
    ]
}